% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIDZN.R
\name{PIDZNclose}
\alias{PIDZNclose}
\title{PID Tuning ZN Closed Loop Tuning}
\usage{
PIDZNclose(Ku, Tu = 0, controltype = 4)
}
\arguments{
\item{:}{Ku, Tu, controltype}
}
\value{
: PID tuning parameters
}
\description{
1. Ziegler Nichols Closed Loop Tuning
Step 1 -  Set integral (I) and derivative (D) to 0
Step 2 -  Increase proportinal gain (P) from 0 until it reaches
          ultimate gain (Ku), i.e. when output of the control loop has
          stable and consistent oscillation
Step 3 -  Ku and oscillation period are used to set PID gains depending
          on the type of controller used
Control Types
1.P                     -   Kp=0.5Ku
2.PI                    -   Kp=0.45Ku;  Ti=Tu/1.2
3.PD                    -   Kp=0.8Ku;               ;Td=Tu/8
4.Classic PID           -   Kp=0.6Ku;   Ti=Tu/2     ;Td=Tu/8
5.Pessen Integral Rule  -   Kp=0.7Ku;   Ti=Tu/2.5   ;Td=3Tu/20
6.Some OVershoot        -   Kp=0.33Ku;  Ti=Tu/2     ;Td=Tu/3
7.No OVershoot          -   Kp=0.2Ku;   Ti=Tu/2     ;Td=Tu/3
u(t)=Kp(e(t) + (1/Ti)int(0,t)(e(T)dT)  + Td*de(t)/dt)
}

