% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIDZN.R
\name{PIDZNopen}
\alias{PIDZNopen}
\title{PID Tuning ZN Open Loop Tuning}
\usage{
PIDZNopen(K, Ts, Td, controltype = 3)
}
\arguments{
\item{:}{K, Ts, Td, controltype}
}
\value{
: PID tuning parameters
}
\description{
2. Ziegler Nichols Open Loop Tuning
Step 1 - Set controller to open loop
Step 2 - Introduce step change in actuator
Step 3 - Determine Steady State Gain K=/(u2-u1) => tangent at point of estimated maximum slope
Step 4 - Determine time delay (Td) => duration from change of u to the point where tangent cuts y1
Step 5 - Determine Time Constant (Ts) => duration where tangent cuts y1 and y2
Control Types
1.P                     -   Kp=Ts/(Td*K)
2.PI                    -   Kp=0.9*Ts/(Td*K);   Ti=3.3*Td
3.PID                   -   Kp=1.2*Ts/(Td*K);   Ti=2.0*Td   ;Td=0.5*Td
}

